Tentar apagar Escola c/ Curso = Status: 400 Bad Requets

+++ src/main/java/com/example/poo_projeto_ac2/controller/EscolaController.java
@@ -4,9 +4,11 @@ import java.util.List;
 
 import javax.servlet.http.HttpServletRequest;
 
+import com.example.poo_projeto_ac2.dto.CursoDTO;
 import com.example.poo_projeto_ac2.dto.EscolaDTO;
 import com.example.poo_projeto_ac2.model.Curso;
 import com.example.poo_projeto_ac2.model.Escola;
+import com.example.poo_projeto_ac2.repository.EscolaRepository;
 import com.example.poo_projeto_ac2.service.CursoService;
 import com.example.poo_projeto_ac2.service.EscolaService;
 
@@ -31,6 +33,9 @@ public class EscolaController {
     private EscolaService escolaService;
 
     @Autowired
+    EscolaRepository escolaRepositorio;
+
+    @Autowired
     private CursoService cursoService;
 
     @GetMapping()
@@ -46,7 +51,7 @@ public class EscolaController {
 
     @PostMapping()
     public ResponseEntity<Void> salvar(@RequestBody EscolaDTO escolaDTO, HttpServletRequest request, UriComponentsBuilder builder){
-        Escola escola = escolaService.saveFromDTO(escolaDTO);
+        Escola escola = escolaService.saveDTO(escolaDTO);
         escola = escolaService.salvar(escola);
         UriComponents uriComponents = builder.path(request.getRequestURI() + "/" + escola.getId()).build();
         return ResponseEntity.created(uriComponents.toUri()).build();
@@ -54,14 +59,19 @@ public class EscolaController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> remover(@PathVariable int id){
+        //Verefica de escola possue curso antes da vereficar confição do 'if'
+        escolaService.validaEscola(escolaRepositorio.getEscolaPorId(id).get());
+        if(escolaRepositorio.getEscolaPorId(id).get().isTemCurso() == false){
+            return ResponseEntity.noContent().build();
+        }
         escolaService.removePorId(id);
-        return ResponseEntity.noContent().build();
+        return ResponseEntity.badRequest().build();
     }
 
     //Unicos dados que poderao ser atualizados são: nome e diretor(a)
     @PutMapping("/{id}")
     public ResponseEntity<Escola> atualizar(@PathVariable int id, @RequestBody EscolaDTO escolaDTO){
-        Escola escola = escolaService.updateFromDTO(escolaDTO);
+        Escola escola = escolaService.updateDTO(escolaDTO);
         escola.setId(id);
         escola = escolaService.atualizar(escola);
         return ResponseEntity.ok(escola);
@@ -75,9 +85,9 @@ public class EscolaController {
     }
 
     @GetMapping("{id}/cursos")
-    public ResponseEntity<List<Curso>> getTodosCurosDaEscola(@PathVariable int id){
+    public List<CursoDTO> getTodosCurosDaEscola(@PathVariable int id) {
         Escola escola = escolaService.getEscolaPorId(id);
-        return ResponseEntity.ok(escola.getCursos());
+        return cursoService.toListDTO(escola.getCursos());
     }
 
 }
